Add files via upload
This is the main module where you will find the routine for encoding a text into its corresponding binary representation. The main encoder terminal as described within the code,   would directly return a dictionary of encodings for each character and the binary string representation of the text upon which the function was called.